refactor(core/directives): separate logic to related modules
