+++ src/core/directives/controller/constants.ts
@@ -1,0 +1,1 @@
+export const REQUIRE_PREFIX_REGEXP = /^(?:(\^\^?)?(\?)?(\^\^?)?)?/;
