+++ src/core/directives/directives_utils.ts
@@ -1,0 +1,38 @@
+import { DirectiveMetadata, ComponentMetadata } from './metadata_directives';
+
+
+export function isAttrDirective( metadata: any ): metadata is DirectiveMetadata {
+  return metadata instanceof DirectiveMetadata && !(metadata instanceof ComponentMetadata);
+}
+export function isComponentDirective( metadata: any ): metadata is ComponentMetadata{
+  return metadata instanceof ComponentMetadata;
+}
+
+/**
+ *
+ * @param scope
+ * @param element
+ * @param ctrl
+ * @param implementsNgOnDestroy
+ * @param watchersToDispose
+ * @param observersToDispose
+ * @private
+ */
+export function _setupDestroyHandler(
+  scope: ng.IScope,
+  element: ng.IAugmentedJQuery,
+  ctrl: any,
+  implementsNgOnDestroy: boolean,
+  watchersToDispose: Function[] = [],
+  observersToDispose: Function[] = []
+): void {
+
+  scope.$on( '$destroy', ()=> {
+
+    if ( implementsNgOnDestroy ) { ctrl.ngOnDestroy() }
+
+    watchersToDispose.forEach( _watcherDispose=>_watcherDispose() );
+    observersToDispose.forEach( _observerDispose=>_observerDispose() );
+    element.off();
+  } );
+}
