+++ src/core/directives/host/constants.ts
@@ -1,0 +1,11 @@
+export type HostBindingsProcessed = {
+  classes: StringMap,
+  attributes: StringMap,
+  properties: StringMap
+}
+export type HostListenersProcessed = {[key:string]:string[]};
+export type HostProcessed = {
+  hostStatic: StringMap,
+  hostBindings: HostBindingsProcessed,
+  hostListeners: HostListenersProcessed
+}
