+++ test/index.ts
@@ -12,7 +12,13 @@ import './core/pipes/pipe_provider.spec';
 import './core/linker/directive_lifecycles_reflector.spec';
 import './core/linker/directive_resolver.spec';
 import './core/directives/directive_provider.spec';
-import './core/directives/util/util.spec';
+import './core/directives/directives_utils.spec';
+import './core/directives/binding/binding_factory.spec';
+import './core/directives/binding/binding_parser.spec';
+import './core/directives/controller/controller_factory.spec';
+import './core/directives/host/host_parser.spec';
+import './core/directives/host/host_resolver.spec';
+import './core/directives/query/children_resolver.spec';
 import './facade/lang.spec';
 import './facade/primitives.spec';
 import './facade/collections.spec';
